Move endpoints for updating information to AdminEndpoint
