+++ src/main/java/no/capraconsulting/endpoints/AdminEndpoint.java
@@ -5,10 +5,12 @@ import com.google.gson.JsonObject;
 import no.capraconsulting.auth.AdminFilter;
 import no.capraconsulting.auth.JwtFilter;
 import no.capraconsulting.client.MSGraphClient;
+import no.capraconsulting.domain.Information;
 import no.capraconsulting.domain.Role;
 import no.capraconsulting.domain.Volunteer;
 import no.capraconsulting.domain.VolunteerRole;
 import no.capraconsulting.repository.AdminRepository;
+import org.json.JSONObject;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -59,5 +61,32 @@ public class AdminEndpoint {
     public Response deleteVolunteer(@PathParam("userId") String userId) {
         AdminRepository.deleteVolunteer(userId);
        return Response.ok().build();
+    }
+
+    @PUT
+    @Path("/information/announcement")
+    @Produces({MediaType.APPLICATION_JSON})
+    public Response updateAnnouncement(String payload) {
+        JSONObject data = new JSONObject(payload);
+        AdminRepository.updateAnnouncement(data.getString("announcement"));
+        return Response.ok().build();
+    }
+
+    @PUT
+    @Path("/information/open")
+    @Produces({MediaType.APPLICATION_JSON})
+    public Response updateIsOpen(String payload) {
+        JSONObject data = new JSONObject(payload);
+        AdminRepository.updateIsOpen(data.getBoolean("isOpen"));
+        return Response.ok(data.toString()).build();
+    }
+
+    @PUT
+    @Path("/information/openinghours")
+    @Produces({MediaType.APPLICATION_JSON})
+    public Response updateOpeningHours(String payload) {
+        Information information = gson.fromJson(payload, Information.class);
+        AdminRepository.updateOpeningHours(information);
+        return Response.ok().build();
     }
 }
