+++ src/main/java/no/capraconsulting/endpoints/OpenLeksehjelpEndpoint.java
@@ -1,9 +1,6 @@
 package no.capraconsulting.endpoints;
 
 import com.google.gson.Gson;
-import no.capraconsulting.auth.JwtFilter;
-import no.capraconsulting.db.Database;
-import no.capraconsulting.domain.Information;
 import no.capraconsulting.utils.EndpointUtils;
 import org.json.JSONObject;
 import org.slf4j.Logger;
@@ -13,7 +10,6 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.sql.SQLException;
-import java.util.*;
 
 @Path("/information")
 public class OpenLeksehjelpEndpoint {
@@ -48,72 +44,4 @@ public class OpenLeksehjelpEndpoint {
             return Response.status(422).build();
         }
     }
-
-    @PUT
-    @JwtFilter
-    @Path("/announcement")
-    @Produces({MediaType.APPLICATION_JSON})
-    public Response updateAnnouncement(String payload) {
-        JSONObject data = new JSONObject(payload);
-
-        try {
-            String insertInformation = "UPDATE INFORMATION SET data=JSON_MODIFY(data, '$.announcement', ?)";
-            Database.INSTANCE.manipulateQuery(insertInformation, false, data.getString("announcement"));
-
-            return Response.status(200).build();
-        } catch (SQLException e) {
-            LOG.error(e.getMessage());
-            return Response.status(422).build();
-        }
-    }
-
-    @PUT
-    @JwtFilter
-    @Path("/open")
-    @Produces({MediaType.APPLICATION_JSON})
-    public Response updateIsOpen(String payload) {
-        JSONObject data = new JSONObject(payload);
-        String insertInformation = "UPDATE INFORMATION SET data=JSON_MODIFY(data, '$.isOpen', ?)";
-
-        try {
-            Database.INSTANCE.manipulateQuery(insertInformation, false, data.getBoolean("isOpen"));
-            return Response.ok(data.toString()).build();
-        } catch (SQLException e) {
-            LOG.error(e.getMessage());
-            return Response.status(422).build();
-        }
-    }
-
-    @PUT
-    @JwtFilter
-    @Path("/openinghours")
-    @Produces({MediaType.APPLICATION_JSON})
-    public Response updateOpeningHours(String payload) {
-        Information information = gson.fromJson(payload, Information.class);
-        List<String> openingDays = Arrays.asList("monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday", "other");
-
-        openingDays.forEach(openingDay -> {
-            try {
-                if (!openingDay.equals("other")) {
-                    String insertStart = "UPDATE INFORMATION SET data=JSON_MODIFY(data, '$."+openingDay+".start', ?)";
-                    String insertEnd = "UPDATE INFORMATION SET data=JSON_MODIFY(data, '$."+openingDay+".end', ?)";
-                    String insertEnabled = "UPDATE INFORMATION SET data=JSON_MODIFY(data, '$."+openingDay+".enabled', ?)";
-
-                    Database.INSTANCE.manipulateQuery(insertStart, false, information.getOpeningHourByDay(openingDay).getStart());
-                    Database.INSTANCE.manipulateQuery(insertEnd, false, information.getOpeningHourByDay(openingDay).getEnd());
-                    Database.INSTANCE.manipulateQuery(insertEnabled, false, information.getOpeningHourByDay(openingDay).isEnabled());
-
-                } else {
-                    String insertMessage = "UPDATE INFORMATION SET data=JSON_MODIFY(data, '$."+openingDay+".message', ?)";
-                    String insertEnabled = "UPDATE INFORMATION SET data=JSON_MODIFY(data, '$."+openingDay+".enabled', ?)";
-                    Database.INSTANCE.manipulateQuery(insertMessage, false, information.getOther().getMessage());
-                    Database.INSTANCE.manipulateQuery(insertEnabled, false, information.getOther().isEnabled());
-                }
-            } catch (SQLException e) {
-                LOG.error(e.getMessage());
-            }
-        });
-        return Response.status(200).build();
-    }
-
 }
